# docs/static/swagger.json
{
...
  "info": {
    "description": "Documenta\u00e7\u00e3o autom\u00e1tica das rotas do sistema com arquitetura DDD e SQL Puro.",
    "title": "Documenta\u00e7\u00e3o da API - Projeto g03",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/scans/{sid}/veiculos": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ve\u00edculos relacionados ao scan"
          },
          "404": {
            "description": "Scan n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lista os ve\u00edculos vinculados a um scan",
        "tags": [
          "Scans"
        ]
      }
    },
...
    "/api/veiculos/": {
      "get": {
        "parameters": [
          {
            "default": 1,
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "in": "query",
            "name": "per_page",
            "type": "integer"
          },
          {
            "in": "query",
            "name": "roubado",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ve\u00edculos"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lista todos os ve\u00edculos com pagina\u00e7\u00e3o e filtro por situa\u00e7\u00e3o de roubo",
        "tags": [
          "Ve\u00edculos"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "cor": {
                  "example": "Branco",
                  "type": "string"
                },
                "data_roubo": {
                  "example": "2024-03-15T10:30:00",
                  "type": "string"
                },
                "modelo": {
                  "example": "Fiat Uno",
                  "type": "string"
                },
                "placa": {
                  "example": "ABC1D23",
                  "type": "string"
                },
                "roubado": {
                  "type": "boolean"
                }
              },
              "required": [
                "placa",
                "modelo",
                "cor",
                "roubado",
                "data_roubo"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ve\u00edculo criado com sucesso"
          },
          "400": {
            "description": "Dados inv\u00e1lidos"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Cadastra um novo ve\u00edculo no sistema",
        "tags": [
          "Ve\u00edculos"
        ]
      }
    },
    "/api/veiculos/placa/{placa}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "placa",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do ve\u00edculo"
          },
          "404": {
            "description": "Ve\u00edculo n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Busca um ve\u00edculo pela placa exata",
        "tags": [
          "Ve\u00edculos"
        ]
      }
    },
    "/api/veiculos/roubados": {
      "get": {
        "responses": {
          "200": {
            "description": "Lista de ve\u00edculos roubados"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lista apenas ve\u00edculos marcados como roubados",
        "tags": [
          "Ve\u00edculos"
        ]
      }
    },
    "/api/veiculos/search": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados da busca"
          },
          "400": {
            "description": "Par\u00e2metro inv\u00e1lido"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Busca ve\u00edculos por placa, modelo ou cor (parcial, case-insensitive)",
        "tags": [
          "Ve\u00edculos"
        ]
      }
    },
    "/api/veiculos/{vid}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "vid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ve\u00edculo removido"
          },
          "404": {
            "description": "Ve\u00edculo n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Remove um ve\u00edculo do sistema",
        "tags": [
          "Ve\u00edculos"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "vid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do ve\u00edculo"
          },
          "404": {
            "description": "Ve\u00edculo n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Busca um ve\u00edculo pelo UUID",
        "tags": [
          "Ve\u00edculos"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "vid",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "cor": {
                  "type": "string"
                },
                "data_roubo": {
                  "type": "string"
                },
                "modelo": {
                  "type": "string"
                },
                "roubado": {
                  "type": "boolean"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ve\u00edculo atualizado"
          },
          "400": {
            "description": "Dados inv\u00e1lidos"
          },
          "404": {
            "description": "Ve\u00edculo n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Atualiza os dados de um ve\u00edculo",
        "tags": [
          "Ve\u00edculos"
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Insira o token JWT no formato: Bearer {seu_token}",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}